vanurses/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
gunicorn>=21.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
stripe>=7.0.0
feedparser>=6.0.0
